\name{learning.test}
\docType{data}
\alias{learning.test}
\title{ Synthetic dataset to test learning algorithms }
\description{

  This a synthetic dataset used as a test case in the
  bnlearn package.

}
\usage{

  learning.test

}
\details{

  The \code{learning.test} dataset containts the following variables:
  \itemize{

    \item \code{A}, a three-level factor with levels \code{a},
            \code{b} and \code{c}.
    \item \code{B}, a three-level factor with levels \code{a},
            \code{b} and \code{c}.
    \item \code{C}, a three-level factor with levels \code{a},
            \code{b} and \code{c}.
    \item \code{D}, a three-level factor with levels \code{a},
            \code{b} and \code{c}.
    \item \code{E}, a three-level factor with levels \code{a},
            \code{b} and \code{c}.
    \item \code{F}, a two-level factor with levels \code{a} 
            and \code{b}.

  }

}
\keyword{datasets}
